styled index page
Moved over the images and converted the LESS from the lineman app over
to SASS.  Basically all you had to do was to change the variable `@` to
`$'.

Also had to place the fonts and the images in the proper place for them
to get sucked up by the asset pipeline.  No longer do you have to add
them to `config/application.rb`, rather they get sucked in
automagically.